requirements.txt
# Module 1
ibm-watsonx-ai==1.4.7
pydantic>=2.0
tenacity>=8.0
requests>=2.28

# Module 2
sentence-transformers>=2.7
transformers>=4.40
torch>=2.2
langchain-chroma>=0.1
langchain-core>=0.2
Pillow>=10.0

# Module 3
openai>=1.0
langchain>=0.2
